backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
python-multipart
aiofiles
celery[redis]
redis
pydantic[email]
python-dotenv
psycopg2-binary
argon2-cffi

